+++ kernel/sysproc.c
@@ -26,7 +26,7 @@ uint64 sys_wait(void) {
 }
 
 uint64 sys_sbrk(void) {
-  int addr;
+  uint64 addr;
   int n;
 
   if (argint(0, &n) < 0)
